+++ source/renderer/command_buffer.cpp
@@ -38,7 +38,7 @@ namespace renderer {
 
         std::vector<VkClearValue> clearValues(clearValueCount);
 
-        for (std::size_t i = 0; i < beginInfo.colourClearValues.size(); i++) {
+        for (std::uint64_t i = 0; i < beginInfo.colourClearValues.size(); i++) {
             auto& vkClearValue = clearValues[i];
             auto& clearValue = beginInfo.colourClearValues[i];
 
@@ -53,7 +53,7 @@ namespace renderer {
         }
 
         if (beginInfo.depthClearValue || beginInfo.stencilClearValue) {
-            std::size_t index = clearValues.size() - 1;
+            std::uint64_t index = clearValues.size() - 1;
 
             auto& vkClearValue = clearValues[index];
 
@@ -108,7 +108,7 @@ namespace renderer {
     void CommandBuffer::copyBuffer(CommandBuffer& commandBuffer, Buffer& source, Buffer& destination, const std::vector<BufferCopyRegion>& copyRegions) {
         std::vector<VkBufferCopy> bufferCopies(copyRegions.size());
 
-        for (std::size_t i = 0; i < bufferCopies.size(); i++) {
+        for (std::uint64_t i = 0; i < bufferCopies.size(); i++) {
             auto& bufferCopy = bufferCopies[i];
             auto& copyRegion = copyRegions[i];
 
@@ -125,7 +125,7 @@ namespace renderer {
     void CommandBuffer::copyBufferToImage(CommandBuffer& commandBuffer, Buffer& source, Image& destination, ImageLayout imageLayout, const std::vector<BufferImageCopyRegion>& copyRegions) {
         std::vector<VkBufferImageCopy> copies(copyRegions.size());
 
-        for (std::size_t i = 0; i < copies.size(); i++) {
+        for (std::uint64_t i = 0; i < copies.size(); i++) {
             auto& copy = copies[i];
             auto& copyRegion = copyRegions[i];
 
@@ -188,7 +188,7 @@ namespace renderer {
     void CommandBuffer::pipelineBarrier(CommandBuffer& commandBuffer, Flags sourcePipelineStage, Flags destinationPipelineStage, const std::vector<ImageMemoryBarrier>& memoryBarriers) {
         std::vector<VkImageMemoryBarrier> barriers(memoryBarriers.size());
 
-        for (std::size_t i = 0; i < barriers.size(); i++) {
+        for (std::uint64_t i = 0; i < barriers.size(); i++) {
 
             VkImageLayout oldLayout;
             VkImageLayout newLayout;
@@ -307,7 +307,7 @@ namespace renderer {
 
         std::vector<VkDescriptorSet> vkSets(sets.size());
 
-        for (std::size_t i = 0; i < vkSets.size(); i++) {
+        for (std::uint64_t i = 0; i < vkSets.size(); i++) {
             vkSets[i] = sets[i].descriptorSet_;
         }
 
@@ -318,7 +318,7 @@ namespace renderer {
         vkCmdBindPipeline(commandBuffer.commandBuffer_, VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline.pipeline_);
     }
 
-    void CommandBuffer::bindVertexBuffers(CommandBuffer& commandBuffer, const std::vector<Buffer>& buffers, const std::vector<std::size_t>& offsets, std::uint32_t first) {
+    void CommandBuffer::bindVertexBuffers(CommandBuffer& commandBuffer, const std::vector<Buffer>& buffers, const std::vector<std::uint64_t>& offsets, std::uint32_t first) {
         std::vector<VkBuffer> vulkanBuffers(buffers.size());
 
         for (std::uint32_t i = 0; i < vulkanBuffers.size(); i++) {
@@ -328,7 +328,7 @@ namespace renderer {
         vkCmdBindVertexBuffers(commandBuffer.commandBuffer_, first, static_cast<std::uint32_t>(vulkanBuffers.size()), vulkanBuffers.data(), offsets.data());
     }
 
-    void CommandBuffer::bindIndexBuffer(CommandBuffer& commandBuffer, Buffer& buffer, std::size_t offset, IndexType indexType) {
+    void CommandBuffer::bindIndexBuffer(CommandBuffer& commandBuffer, Buffer& buffer, std::uint64_t offset, IndexType indexType) {
         VkIndexType type;
 
         switch (indexType) {
